requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
PyMuPDF==1.23.14
python-docx==1.1.0
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0